backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pgvector
sentence-transformers
google-generativeai
requests
pyjwt